Finalise Perf tests
